Flytt typing av hovedytelse til fellesfiler for gjenbruk
