+++ src/main/kotlin/no/nav/tilleggsstonader/soknad/soknad/barnetilsyn/SøknadBarnetilsynDto.kt
@@ -1,17 +1,13 @@
 package no.nav.tilleggsstonader.soknad.soknad.barnetilsyn
 
-import no.nav.tilleggsstonader.kontrakter.felles.Hovedytelse
-import no.nav.tilleggsstonader.kontrakter.søknad.DatoFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.DokumentasjonFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.EnumFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.EnumFlereValgFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.JaNei
-import no.nav.tilleggsstonader.kontrakter.søknad.SelectFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.AnnenAktivitetType
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.TypeBarnepass
-import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.TypePengestøtte
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.ÅrsakBarnepass
-import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.ÅrsakOppholdUtenforNorge
+import no.nav.tilleggsstonader.soknad.soknad.HovedytelseDto
 import no.nav.tilleggsstonader.soknad.soknad.SøknadValideringException
 
 data class SøknadBarnetilsynDto(
@@ -19,29 +15,6 @@ data class SøknadBarnetilsynDto(
     val aktivitet: Aktivitet,
     val barnMedBarnepass: List<BarnMedBarnepass>,
     val dokumentasjon: List<DokumentasjonFelt>,
-)
-
-data class HovedytelseDto(
-    val ytelse: EnumFlereValgFelt<Hovedytelse>,
-    val arbeidOgOpphold: ArbeidOgOppholdDto,
-)
-
-data class ArbeidOgOppholdDto(
-    val jobberIAnnetLand: EnumFelt<JaNei>?,
-    val jobbAnnetLand: SelectFelt<String>?,
-    val harPengestøtteAnnetLand: EnumFlereValgFelt<TypePengestøtte>?,
-    val pengestøtteAnnetLand: SelectFelt<String>?,
-    val harOppholdUtenforNorgeSiste12mnd: EnumFelt<JaNei>?,
-    val oppholdUtenforNorgeSiste12mnd: List<OppholdUtenforNorgeDto>,
-    val harOppholdUtenforNorgeNeste12mnd: EnumFelt<JaNei>?,
-    val oppholdUtenforNorgeNeste12mnd: List<OppholdUtenforNorgeDto>,
-)
-
-data class OppholdUtenforNorgeDto(
-    val land: SelectFelt<String>,
-    val årsak: EnumFlereValgFelt<ÅrsakOppholdUtenforNorge>,
-    val fom: DatoFelt,
-    val tom: DatoFelt,
 )
 
 data class Aktivitet(
